prefer setter annotations over getter annotations
As all of the mappers deal with mapping data onto beans (which they use the setters for),
they should also prefer the annotations found on the setter over the annotations on the getter.

This PR also splits the very unwieldy and confusing BeanMapperTest into three separate tests,
one focusing on general bean mapper properties, one for the PropagateNullable tests and one for
Nested tests.
